fixed warning

+++ UnitTests/Basics/string-utf8-normalize-test.cpp
@@ -168,7 +168,7 @@ BOOST_AUTO_TEST_CASE (tst_2) {
   BOOST_CHECK_EQUAL(expectString, resultString);
   
   len = 0;
-  result = TRI_tolower_utf8(TRI_CORE_MEM_ZONE, (const char*) gruessgott2, strlen((const char*) gruessgott2), &len);
+  result = TRI_tolower_utf8(TRI_CORE_MEM_ZONE, (const char*) gruessgott2, (int32_t) strlen((const char*) gruessgott2), &len);
   //printf("\nOriginal: %s\nLower: %s (%d)\n", gruessgott2, result, len);
   l2 = strlen(result);
   BOOST_CHECK_EQUAL(l1, l2);
